use API to fetch download URL of stable release
looking at cursor.com/downloads we can see that the web app loads a JSON
from https://www.cursor.com/api/download with respective parameters to
derive the actual URL to use for downloading the stable version of
Cursor. this changeset builds on that in order to download the latest
stable release of Cursor, which is currently the 0.47.x series.